strands-agent/pyproject.toml
[project]
name = "strands-venice-agent"
version = "0.3.0"
description = "Venice GLM-4.7 uncensored research agent — Strands Agents SDK"
requires-python = ">=3.10"
dependencies = [
    "strands-agents",
    "strands-agents-tools",
    "strands-agents-builder",
    "strands-deep-agents",
    "openai",
    "python-dotenv",
    "fastapi[standard]",
    "uvicorn[standard]",
    "httpx",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "agent", "config", "tools", "prompts", "thought_refiner"]

[tool.setuptools.packages.find]
include = ["plugins*", "evals*"]

[tool.pytest.ini_options]
testpaths = ["evals"]
asyncio_mode = "auto"
markers = [
    "integ: integration tests requiring API keys",
]
